+++ store/fuzz_test.go
@@ -3,8 +3,10 @@ package store
 import (
 	"bytes"
 	"crypto/rand"
-	"encoding/hex"
+	"fmt"
+	"github.com/dgraph-io/badger/v4"
 	"github.com/ginchuco/ginchu/types"
+	"github.com/stretchr/testify/require"
 	math "math/rand"
 	"testing"
 )
@@ -15,22 +17,23 @@ const (
 	SetTesting TestingOp = iota
 	DelTesting
 	GetTesting
-	WriteTesting
 	IterateTesting
-	Commit
 )
 
 func TestFuzz(t *testing.T) {
-	logger := types.NewDefaultLogger()
-	cacheStore := NewStoreCache(NewMemoryStore(logger))
+	db, err := badger.OpenManaged(badger.DefaultOptions("").
+		WithInMemory(true).WithLoggingLevel(badger.ERROR))
+	require.NoError(t, err)
+	store, _, cleanup := testStore(t)
+	defer cleanup()
 	keys := make([][]byte, 0)
-	compareStore := NewMemoryStore(logger)
+	compareStore2 := NewTxnWrapper(db.NewTransactionAt(1, true), types.NewDefaultLogger(), stateStorePrefix)
 	for i := 0; i < 1000; i++ {
-		doRandomOperation(t, cacheStore, compareStore, keys)
+		doRandomOperation(t, store, compareStore2, keys)
 	}
 }
 
-func doRandomOperation(t *testing.T, db *storeCache, compare types.StoreI, keys [][]byte) {
+func doRandomOperation(t *testing.T, db types.KVStoreI, compare types.KVStoreI, keys [][]byte) {
 	k, v := getRandomBytes(t), getRandomBytes(t)
 	switch getRandomOperation(t) {
 	case SetTesting:
@@ -45,29 +48,29 @@ func doRandomOperation(t *testing.T, db *storeCache, compare types.StoreI, keys 
 		k = randomTestKey(t, k, keys)
 		v1, v2 := testDBGet(t, db, k), testDBGet(t, compare, k)
 		if !bytes.Equal(v1, v2) {
-			t.Fatalf("db.Get=%s compare.Get=%s", v1, v2)
+			fmt.Printf("key=%s db.Get=%s compare.Get=%s\n", k, v1, v2)
 		}
-	case WriteTesting:
-		db.Write()
+		require.Equalf(t, v1, v2, "key=%s db.Get=%s compare.Get=%s", k, v1, v2)
 	case IterateTesting:
-		testCompareIterators(t, db, compare)
-	case Commit:
 		testCompareIterators(t, db, compare)
 	default:
 		t.Fatal("invalid op")
 	}
 }
 
-func getRandomBytes(_ *testing.T) []byte {
-	bz := make([]byte, math.Intn(1000))
+func getRandomBytes(t *testing.T) []byte {
+	bz := make([]byte, math.Intn(999)+40)
 	if _, err := rand.Read(bz); err != nil {
-		return nil
+		t.Fatal(err)
+	}
+	if bz == nil {
+		t.Fatal("nil bz")
 	}
 	return bz
 }
 
 func getRandomOperation(_ *testing.T) TestingOp {
-	return TestingOp(math.Intn(3))
+	return TestingOp(math.Intn(4))
 }
 
 func randomTestKey(_ *testing.T, k []byte, keys [][]byte) []byte {
@@ -79,65 +82,41 @@ func randomTestKey(_ *testing.T, k []byte, keys [][]byte) []byte {
 	return k
 }
 
-func testDBSet(t *testing.T, db types.StoreI, k, v []byte) {
-	if err := db.Set(k, v); err != nil {
-		t.Fatal(err)
-	}
+func testDBSet(t *testing.T, db types.WritableStoreI, k, v []byte) {
+	require.NoError(t, db.Set(k, v))
 }
 
-func testDBDelete(t *testing.T, db types.StoreI, k []byte) {
-	if err := db.Delete(k); err != nil {
-		t.Fatal(err)
-	}
+func testDBDelete(t *testing.T, db types.WritableStoreI, k []byte) {
+	require.NoError(t, db.Delete(k))
 }
 
-func testDBGet(t *testing.T, db types.StoreI, k []byte) (value []byte) {
+func testDBGet(t *testing.T, db types.KVStoreI, k []byte) (value []byte) {
 	value, err := db.Get(k)
-	if err != nil {
-		t.Fatal(err)
-	}
+	require.NoError(t, err)
 	return
 }
 
-func testCommit(t *testing.T, db *storeCache, compare types.StoreI) {
-	db.Write()
-}
-
-func testCompareIterators(t *testing.T, db *storeCache, compare types.StoreI) {
+func testCompareIterators(t *testing.T, db types.KVStoreI, compare types.KVStoreI) {
 	var (
 		it1, it2 types.IteratorI
 		err      error
 	)
 	switch math.Intn(1) {
 	case 0:
-		it1, err = db.Iterator(nil, nil)
-		if err != nil {
-			t.Fatal(err)
-		}
-		it2, err = compare.Iterator(nil, nil)
-		if err != nil {
-			t.Fatal(err)
-		}
+		it1, err = db.Iterator(nil)
+		require.NoError(t, err)
+		it2, err = compare.Iterator(nil)
+		require.NoError(t, err)
 	case 1:
-		it1, err = db.ReverseIterator(nil, nil)
-		if err != nil {
-			t.Fatal(err)
-		}
-		it2, err = compare.ReverseIterator(nil, nil)
-		if err != nil {
-			t.Fatal(err)
-		}
+		it1, err = db.RevIterator(nil)
+		require.NoError(t, err)
+		it2, err = compare.RevIterator(nil)
+		require.NoError(t, err)
 	}
 	defer func() { it1.Close(); it2.Close() }()
 	for ; func() bool { return it1.Valid() || it2.Valid() }(); func() { it1.Next(); it2.Next() }() {
-		if it1.Valid() != it2.Valid() {
-			t.Fatalf("it1.valid=%t it2.valid=%t ", it1.Valid(), it2.Valid())
-		}
-		if bytes.Equal(it1.Key(), it2.Key()) {
-			t.Fatalf("it1.key=%s it2.key=%s ", hex.EncodeToString(it1.Key()), hex.EncodeToString(it2.Key()))
-		}
-		if bytes.Equal(it1.Value(), it2.Value()) {
-			t.Fatalf("it1.key=%s it2.key=%s ", hex.EncodeToString(it1.Key()), hex.EncodeToString(it2.Key()))
-		}
+		require.Equal(t, it1.Valid(), it2.Valid(), fmt.Sprintf("it1.valid=%t it2.valid=%t ", it1.Valid(), it2.Valid()))
+		require.Equal(t, it1.Key(), it2.Key(), fmt.Sprintf("it1.key=%s it2.key=%s ", it1.Key(), it2.Key()))
+		require.Equal(t, it1.Value(), it2.Value(), fmt.Sprintf("it1.value=%s it2.value=%s ", it1.Value(), it2.Value()))
 	}
 }