+++ store/store.go
@@ -1,131 +1,171 @@
 package store
 
 import (
-	"github.com/ginchuco/ginchu/crypto"
-	"github.com/ginchuco/ginchu/store/smt"
+	"fmt"
+	"github.com/dgraph-io/badger/v4"
+	"github.com/ginchuco/ginchu/codec"
 	"github.com/ginchuco/ginchu/types"
-	"github.com/syndtr/goleveldb/leveldb"
-	"github.com/syndtr/goleveldb/leveldb/opt"
-	"github.com/syndtr/goleveldb/leveldb/storage"
+	"math"
 )
 
-var _ types.CommitStoreI = &Store{}
+const (
+	stateStorePrefix      = "s/"
+	stateCommitmentPrefix = "c/"
+	stateCommitIDPrefix   = "x/"
+)
+
+var (
+	cdc        = codec.Protobuf{}
+	lastCIDKey = []byte("xl/")
+)
 
 type Store struct {
-	log   types.LoggerI
-	store *storeCache
-	smt   *smt.SMT
+	version uint64
+	log     types.LoggerI
+	db      *badger.DB
+	writer  *badger.Txn
+	ss      *TxnWrapper
+	sc      *SMTWrapper
 }
 
-func NewStore(path string, log types.LoggerI, opts *opt.Options) types.CommitStoreI {
-	db, err := leveldb.OpenFile(path, opts)
+func NewStore(path string, log types.LoggerI) (*Store, error) {
+	db, err := badger.OpenManaged(badger.DefaultOptions(path).
+		WithLoggingLevel(badger.ERROR))
 	if err != nil {
-		log.Fatal(newOpenStoreError(err).Error())
+		return nil, err
 	}
 	return newStore(db, log)
 }
 
-func NewMemoryStore(log types.LoggerI) types.CommitStoreI {
-	db, err := leveldb.Open(storage.NewMemStorage(), nil)
+func NewStoreInMemory(log types.LoggerI) (*Store, error) {
+	db, err := badger.OpenManaged(badger.DefaultOptions("").
+		WithInMemory(true).WithLoggingLevel(badger.ERROR))
 	if err != nil {
-		log.Fatal(newOpenStoreError(err).Error())
+		return nil, err
 	}
 	return newStore(db, log)
 }
 
-func newStore(db *leveldb.DB, log types.LoggerI) types.CommitStoreI {
-	levelDB := NewStoreCache(&LevelDBWrapper{db})
-	return &Store{log, levelDB, smt.NewSMT(levelDB, crypto.Hasher())}
+func newStore(db *badger.DB, log types.LoggerI) (*Store, error) {
+	id := getLatestCommitID(db, log)
+	writer := db.NewTransactionAt(uint64(id.Height), true)
+	return &Store{
+		version: uint64(id.Height),
+		log:     log,
+		db:      db,
+		writer:  writer,
+		ss:      NewTxnWrapper(writer, log, stateStorePrefix),
+		sc:      NewSMTWrapper(NewTxnWrapper(writer, log, stateCommitmentPrefix), id.Root, log),
+	}, nil
 }
 
-func (s *Store) Get(key []byte) ([]byte, error)                { return s.store.Get(key) }
-func (s *Store) Iterator(b, e []byte) (types.IteratorI, error) { return s.store.Iterator(b, e) }
-func (s *Store) ReverseIterator(b, e []byte) (types.IteratorI, error) {
-	return s.store.ReverseIterator(b, e)
+func (s *Store) NewReadOnly(version uint64) (*Store, error) {
+	id, err := s.getCommitID(version)
+	if err != nil {
+		return nil, err
+	}
+	reader := s.db.NewTransactionAt(version, false)
+	return &Store{
+		version: version,
+		log:     nil,
+		ss:      NewTxnWrapper(reader, s.log, stateStorePrefix),
+		sc:      NewSMTWrapper(NewTxnWrapper(reader, s.log, stateCommitmentPrefix), id.Root, s.log),
+	}, nil
 }
 
-func (s *Store) Set(key []byte, value []byte) error {
-	if err := s.store.Set(key, value); err != nil {
+func (s *Store) Get(key []byte) ([]byte, error) { return s.ss.Get(key) }
+func (s *Store) Set(k, v []byte) error {
+	if err := s.ss.Set(k, v); err != nil {
 		return err
 	}
-	if err := s.smt.Update(crypto.Hash(key), value); err != nil {
-		return err
-	}
-	return nil
+	return s.sc.Set(k, v)
 }
 
-func (s *Store) Delete(key []byte) error {
-	if err := s.store.Delete(key); err != nil {
+func (s *Store) Delete(k []byte) error {
+	if err := s.ss.Delete(k); err != nil {
 		return err
 	}
-	if err := s.smt.Delete(crypto.Hash(key)); err != nil && err != smt.ErrKeyNotPresent {
-		return err
-	}
-	return nil
+	return s.sc.Delete(k)
 }
 
-func (s *Store) GetProof(key []byte) (compact, value []byte, err error) {
-	value, err = s.store.Get(key)
+func (s *Store) GetProof(k []byte) ([]byte, []byte, error) {
+	val, err := s.ss.Get(k)
 	if err != nil {
 		return nil, nil, err
 	}
-	proof, err := s.smt.Prove(crypto.Hash(key))
+	proof, _, err := s.sc.GetProof(k)
+	return proof, val, err
+}
+
+func (s *Store) VerifyProof(k, v, p []byte) bool               { return s.sc.VerifyProof(k, v, p) }
+func (s *Store) Iterator(p []byte) (types.IteratorI, error)    { return s.ss.Iterator(p) }
+func (s *Store) RevIterator(p []byte) (types.IteratorI, error) { return s.ss.RevIterator(p) }
+
+func (s *Store) Commit() (root []byte, err error) {
+	s.version++
+	root, err = s.sc.Commit()
 	if err != nil {
-		return nil, nil, err
+		return nil, err
 	}
-	cProof, err := smt.CompactProof(proof, &s.smt.BaseSMT)
+	if err = s.setCommitID(s.version, root); err != nil {
+		return nil, err
+	}
+	if err = s.writer.CommitAt(s.version, nil); err != nil {
+		return nil, err
+	}
+	s.writer = s.db.NewTransactionAt(s.version, true)
+	s.ss.parent = s.writer
+	s.sc.SetParent(NewTxnWrapper(s.writer, s.log, stateCommitmentPrefix), root)
+	return
+}
+
+func (s *Store) commitIDKey(version uint64) []byte {
+	return []byte(fmt.Sprintf("%s%d", stateCommitIDPrefix, version))
+}
+
+func (s *Store) getCommitID(version uint64) (id CommitID, err error) {
+	var bz []byte
+	bz, err = NewTxnWrapper(s.writer, s.log, "").Get(s.commitIDKey(version))
 	if err != nil {
-		return nil, nil, err
+		return
 	}
-	ptr := &SparseCompactMerkleProof{
-		SideNodes:             cProof.SideNodes,
-		NonMembershipLeafData: cProof.NonMembershipLeafData,
-		BitMask:               cProof.BitMask,
-		NumSideNodes:          uint32(cProof.NumSideNodes),
-		SiblingData:           cProof.SiblingData,
-	}
-	if compact, err = cdc.Marshal(ptr); err != nil {
-		return nil, nil, err
+	if err = cdc.Unmarshal(bz, &id); err != nil {
+		return
 	}
 	return
 }
 
-func (s *Store) VerifyProof(key, value, proof []byte) bool {
-	ptr := &SparseCompactMerkleProof{}
-	if err := cdc.Unmarshal(proof, ptr); err != nil {
-		s.log.Error(newUnmarshalCompactProofError(err).Error())
-		return false
+func (s *Store) setCommitID(version uint64, root []byte) error {
+	w := NewTxnWrapper(s.writer, s.log, "")
+	value, err := cdc.Marshal(&CommitID{
+		Height: uint32(version),
+		Root:   root,
+	})
+	if err = w.Set(lastCIDKey, value); err != nil {
+		return err
 	}
-	sparseMerkleProof, err := smt.DecompactProof(smt.SparseCompactMerkleProof{
-		SideNodes:             ptr.SideNodes,
-		NonMembershipLeafData: ptr.NonMembershipLeafData,
-		BitMask:               ptr.BitMask,
-		NumSideNodes:          int(ptr.NumSideNodes),
-		SiblingData:           ptr.SiblingData,
-	}, &s.smt.BaseSMT)
+	key := s.commitIDKey(version)
 	if err != nil {
-		s.log.Error(newDecompactProofError(err).Error())
-		return false
+		return err
 	}
-	return smt.VerifyProof(sparseMerkleProof, s.smt.Root(), crypto.Hash(key), value, &s.smt.BaseSMT)
+	return w.Set(key, value)
 }
 
-func (s *Store) Commit() (root []byte, err error) {
-	// sm.Commit() writes the tree to the cachedStore
-	if err = s.smt.Commit(); err != nil {
-		return nil, err
+func getLatestCommitID(db *badger.DB, log types.LoggerI) (id *CommitID) {
+	tx := NewTxnWrapper(db.NewTransactionAt(math.MaxUint64, false), log, "")
+	defer tx.Close()
+	id = new(CommitID)
+	bz, err := tx.Get(lastCIDKey)
+	if err != nil {
+		log.Fatalf("getLatestCommitID() failed with err: %s", err.Error())
 	}
-	// store.Write() writes the query data and the
-	// smt data to the persisted database
-	s.store.Write()
-	// return copy of root
-	rootCpy := make([]byte, len(s.smt.SavedRoot))
-	copy(rootCpy, s.smt.SavedRoot)
-	return rootCpy, err
+	if err = cdc.Unmarshal(bz, id); err != nil {
+		log.Fatalf("unmarshalCommitID() failed with err: %s", err.Error())
+	}
+	return
 }
 
-func (s *Store) CacheWrap() types.StoreI {
-	return NewStoreCache(s)
+func (s *Store) Close() error {
+	s.writer.Discard()
+	return s.db.Close()
 }
-
-func (s *Store) Close() error { return s.store.parent.Close() }